Move JSONSchema related functions to a service
